added pagination and delete option
